Move time-invariant fields to a separate input stream
In order to avoid writing time-invariant fields to every restart file, we define
a new input stream for MPAS-Atmosphere, the "static" stream. With this new stream,
static fields can be removed from the "restart" stream. When the model starts up,
we now first read the "static" stream, then we read either the "input" or "restart"
stream, depending on whether the simulation is a restart run or not.

There are several fields worth noting:

- The dss, meshScalingDel{2,4}, east, and north fields are always computed when
  the model starts up, so we never need to read these

- The pin and ozmixm fields have been moved to the "static" stream, but these
  are only available after the model starts up and reads in the OZONE_DAT.TBL
  and other files. Some cleverness is needed to get these fields into the file
  used for the "static" stream.

Also, because the model infrastructure needs mesh fields for the boostrapping
process, we modify the MPAS-A core interface to specify the "static" stream as
the stream that contains mesh fields for bootstrapping, regardless of whether
we are doing a restart or a cold-start run.